+++ shellfoundry/commands/install_command.py
@@ -1,12 +1,17 @@
-import os
-from urllib2 import HTTPError, URLError
+# !/usr/bin/python
+# -*- coding: utf-8 -*-
+
 import click
-from shellfoundry.utilities.shell_config_reader import ShellConfigReader
+import os
+
+from urllib2 import HTTPError, URLError
+
+from shellfoundry.exceptions import FatalError
 from shellfoundry.utilities.config_reader import Configuration, CloudShellConfigReader
 from shellfoundry.utilities.installer import ShellInstaller
+from shellfoundry.utilities.shell_config_reader import ShellConfigReader
 from shellfoundry.utilities.shell_package import ShellPackage
 from shellfoundry.utilities.shell_package_installer import ShellPackageInstaller
-from shellfoundry.exceptions import FatalError
 
 
 class InstallCommandExecutor(object):
@@ -20,11 +25,15 @@ class InstallCommandExecutor(object):
     def install(self):
         current_path = os.getcwd()
         shell_package = ShellPackage(current_path)
-        if shell_package.is_tosca():
-            self.shell_package_installer.install(current_path)
+        if shell_package.is_layer_one():
+            click.secho("Installing a L1 shell directly via shellfoundry is not supported. "
+                        "Please follow the L1 shell import procedure described in help.quali.com.", fg="yellow")
         else:
-            self._install_old_school_shell()
-        click.secho('Successfully installed shell', fg='green')
+            if shell_package.is_tosca():
+                self.shell_package_installer.install(current_path)
+            else:
+                self._install_old_school_shell()
+            click.secho('Successfully installed shell', fg='green')
 
     def _install_old_school_shell(self):
         error = None