+++ shellfoundry/utilities/config/config_providers.py
@@ -12,6 +12,7 @@ class LocalConfigProvider(object):
             click.echo('Using local configuration...')
         return path
 
+
 class GlobalConfigProvider(object):
     QUALI = 'Quali'
     PRODUCT = 'shellfoundry'
@@ -20,6 +21,7 @@ class GlobalConfigProvider(object):
         sf_name = os.path.join(GlobalConfigProvider.QUALI, GlobalConfigProvider.PRODUCT)
         app_dir_path = click.get_app_dir(sf_name)
         return os.path.join(app_dir_path, GLOBAL_CONFIG_NAME)
+
 
 class ConfigProvider(object):
     def __init__(self, *args):
@@ -31,9 +33,10 @@ class ConfigProvider(object):
             if os.path.exists(config_path):
                 return config_path
 
+
 class DefaultConfigProvider(ConfigProvider):
     def __init__(self):
-        ConfigProvider.__init__(self, (GlobalConfigProvider())) # The order do matters
+        ConfigProvider.__init__(self, (GlobalConfigProvider()))  # The order do matters
         self.default_provider = LocalConfigProvider()
         # self.fallback_provider = LocalConfigProvider()
 
@@ -41,4 +44,4 @@ class DefaultConfigProvider(ConfigProvider):
         config_path = self.default_provider.get_config_path()
         if not os.path.exists(config_path):
             config_path = ConfigProvider.get_config_path(self)
-        return config_path #or self.fallback_provider.get_config_path()
+        return config_path  # or self.fallback_provider.get_config_path()