+++ shellfoundry/utilities/cookiecutter_integration.py
@@ -1,14 +1,21 @@
-from cookiecutter.main import cookiecutter
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
 import os
 
+from cookiecutter.main import cookiecutter
+from shellfoundry.utilities.config_reader import Configuration, CloudShellConfigReader
 
-class CookiecutterTemplateCompiler:
+
+class CookiecutterTemplateCompiler(object):
     def __init__(self):
-        pass
+        self.cloudshell_config_reader = Configuration(CloudShellConfigReader())
 
     def compile_template(self, shell_name, template_path, extra_context, running_on_same_folder):
 
-        extra_context['project_name'] = shell_name
+        extra_context["project_name"] = shell_name
+        extra_context["full_name"] = self.cloudshell_config_reader.read().author
+
         if running_on_same_folder:
             output_dir = os.path.pardir
         else: