+++ tests/test_commands/test_install_command.py
@@ -30,6 +30,7 @@ install:
     username: YOUR_USERNAME
     password: YOUR_PASSWORD
     domain: Global
+    author: AUTHOR
     """)
         os.chdir('nut_shell')
 
@@ -41,7 +42,7 @@ install:
 
         # Assert
         mock_installer.install.assert_called_once_with('nut_shell', InstallConfig('localhost', 9000, 'YOUR_USERNAME',
-                                                                                  'YOUR_PASSWORD', 'Global'))
+                                                                                  'YOUR_PASSWORD', 'Global', "AUTHOR"))
         secho_mock.assert_any_call('Successfully installed shell', fg='green')
 
     def test_tosca_based_shell_installed_when_tosca_meta_file_exists(self):
@@ -62,6 +63,23 @@ install:
 
         # Assert
         self.assertTrue(mock_shell_package_installer.install.called)
+
+    @patch('click.secho')
+    def test_install_layer_one_shell(self, secho_mock):
+        # Arrange
+        self.fs.CreateFile('cloudshell-L1-test/datamodel/datamodel.xml')
+        os.chdir('cloudshell-L1-test')
+
+        mock_shell_package_installer = MagicMock()
+        command_executor = InstallCommandExecutor(shell_package_installer=mock_shell_package_installer)
+
+        # Act
+        command_executor.install()
+
+        # Assert
+        secho_mock.assert_any_call("Installing a L1 shell directly via shellfoundry is not supported. "
+                                   "Please follow the L1 shell import procedure described in help.quali.com.",
+                                   fg="yellow")
 
     def test_proper_error_message_displayed_when_login_failed(self):
         # Arrange